requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
sqlalchemy[asyncio]>=2.0.30
asyncpg>=0.29.0
alembic>=1.13.1
pydantic-settings>=2.2.1
bcrypt>=4.0.0
python-jose[cryptography]>=3.3.0
playwright>=1.44.0
aiohttp>=3.9.5
stripe>=9.0.0
jinja2>=3.1.4
python-multipart>=0.0.9
slowapi>=0.1.9
structlog>=24.2.0
pytest>=8.2.0
pytest-asyncio>=0.23.7
httpx>=0.27.0
redis>=5.0.4
itsdangerous>=2.2.0
